configured the weather widget for geolocation
