+++ src/index.js
@@ -12,12 +12,16 @@ import geolocation from './js/geolocationService';
 
 favorites.loader(); // получаем данные при загрузке страницы из localStorage
 
-apiService.fetchTodayWeather().then(() => {
-  const widgetMarkup = widgetTemplate(apiService.todayResponse);
-  refs.currentWeather.innerHTML = widgetMarkup;
-  createClock('#timer-1');
-  renderSunsetTime(apiService.todayResponse);
-});
+// ============
+// перенес в геолокацию при отмене
+
+// apiService.fetchTodayWeather().then(() => {
+//   const widgetMarkup = widgetTemplate(apiService.todayResponse);
+//   refs.currentWeather.innerHTML = widgetMarkup;
+//   createClock('#timer-1');
+//   renderSunsetTime(apiService.todayResponse);
+// });
+// ============
 
 // refs.switchDaysBtn.addEventListener('click', async event => {
 //   event.preventDefault();
@@ -39,6 +43,7 @@ refs.searchForm.addEventListener('submit', async event => {
     refs.currentWeather.innerHTML = widgetMarkup;
     createClock('#timer-1');
     renderSunsetTime(apiService.todayResponse);
+
     backgroundImageService.background(refs.formInput.value);
   } catch (error) {
     console.log(error);
