+++ src/js/geolocationService.js
@@ -2,15 +2,16 @@ import { alert } from '@pnotify/core';
 import * as PNotifyConfirm from '@pnotify/confirm';
 import { defaults } from '@pnotify/core';
 import axios from 'axios';
-
-import backgroundImageService from './backgroundService';
 import apiService from './apiService';
 import widgetTemplate from '../templates/current-weather.hbs';
 import refs from './refs';
+import createClock from './timerService';
+import renderSunsetTime from './render-sunset-time';
+import backgroundImageService from './backgroundService';
 
 defaults.width = '300px';
 
-let geolocation = '';
+let geolocation = 'kyiv';
 
 const notice = alert({
   title: 'Know your location',
@@ -29,10 +30,17 @@ const notice = alert({
   ]),
 });
 notice.on('pnotify:confirm', () => {
-  navigator.geolocation.getCurrentPosition(success);
+  navigator.geolocation.getCurrentPosition(success, error);
 });
 notice.on('pnotify:cancel', () => {
-  backgroundImageService.background('kyiv');
+  apiService.fetchTodayWeather().then(() => {
+    const widgetMarkup = widgetTemplate(apiService.todayResponse);
+    refs.currentWeather.innerHTML = widgetMarkup;
+    createClock('#timer-1');
+    renderSunsetTime(apiService.todayResponse);
+  });
+
+  backgroundImageService.background(geolocation);
 });
 
 function success(position) {
@@ -45,8 +53,25 @@ function success(position) {
     )
     .then(res => {
       geolocation = res.data.results[0].components.city;
+
+      apiWidget(geolocation);
+
       backgroundImageService.background(geolocation);
     });
+}
+
+function error(error) {
+  console.log(`not found your place, error${error}`);
+}
+
+function apiWidget(place) {
+  apiService.query = place.toLowerCase();
+  apiService.fetchTodayWeather().then(() => {
+    const widgetMarkup = widgetTemplate(apiService.todayResponse);
+    refs.currentWeather.innerHTML = widgetMarkup;
+    createClock('#timer-1');
+    renderSunsetTime(apiService.todayResponse);
+  });
 }
 
 export default notice;
